+++ src/views/Dashboard.vue
@@ -186,7 +186,7 @@ export default {
   data() {
     return {
       loading: false,
-      version: "0.1.3",
+      version: "0.1.4",
       labels: ["生物", "物品", "物体", "任务", "对话", "内建脚本", "技能"],
     };
   },